# frontend/src/api/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "title": "costrict",
    "description": "costrict api",
    "version": "0.0.3"
  },
  "paths": {
...
    "/session/{sessionID}/message/{messageID}": {
      "get": {
        "operationId": "session.message",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "path",
            "name": "sessionID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Session ID"
          },
          {
            "in": "path",
            "name": "messageID",
            "schema": {
              "type": "string"
            },
            "required": true,
            "description": "Message ID"
          }
        ],
        "summary": "Get message",
        "description": "Retrieve a specific message from a session by its message ID.",
        "responses": {
          "200": {
            "description": "Message",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "info": {
                      "$ref": "#/components/schemas/Message"
                    },
                    "parts": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Part"
                      }
                    }
                  },
                  "required": [
                    "info",
                    "parts"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BadRequestError"
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotFoundError"
                }
              }
            }
          }
        }
      }
    },
...
    "/tui/control/response": {
      "post": {
        "operationId": "tui.control.response",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Submit TUI response",
        "description": "Submit a response to the TUI request queue to complete a pending request.",
        "responses": {
          "200": {
            "description": "Response submitted successfully",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          }
        }
      }
    },
...
    "/event": {
      "get": {
        "operationId": "event.subscribe",
        "parameters": [
          {
            "in": "query",
            "name": "directory",
            "schema": {
              "type": "string"
            }
          }
        ],
        "summary": "Subscribe to events",
        "description": "Get events",
        "responses": {
          "200": {
            "description": "Event stream",
            "content": {
              "text/event-stream": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "UserMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "const": "user"
          },
          "time": {
            "type": "object",
            "properties": {
              "created": {
                "type": "number"
              }
            },
            "required": [
              "created"
            ]
          },
          "summary": {
            "type": "object",
            "properties": {
              "title": {
                "type": "string"
              },
              "body": {
                "type": "string"
              },
              "diffs": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/FileDiff"
                }
              }
            },
            "required": [
              "diffs"
            ]
          },
          "agent": {
            "type": "string"
          },
          "model": {
            "type": "object",
            "properties": {
              "providerID": {
                "type": "string"
              },
              "modelID": {
                "type": "string"
              }
            },
            "required": [
              "providerID",
              "modelID"
            ]
          },
          "system": {
            "type": "string"
          },
          "tools": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "boolean"
            }
          },
          "variant": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "role",
          "time",
          "agent",
          "model"
        ]
      },
...
      "AssistantMessage": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "role": {
            "type": "string",
            "const": "assistant"
          },
          "time": {
            "type": "object",
            "properties": {
              "created": {
                "type": "number"
              },
              "completed": {
                "type": "number"
              }
            },
            "required": [
              "created"
            ]
          },
          "error": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ProviderAuthError"
              },
              {
                "$ref": "#/components/schemas/UnknownError"
              },
              {
                "$ref": "#/components/schemas/MessageOutputLengthError"
              },
              {
                "$ref": "#/components/schemas/MessageAbortedError"
              },
              {
                "$ref": "#/components/schemas/APIError"
              }
            ]
          },
          "parentID": {
            "type": "string"
          },
          "modelID": {
            "type": "string"
          },
          "providerID": {
            "type": "string"
          },
          "mode": {
            "type": "string"
          },
          "agent": {
            "type": "string"
          },
          "path": {
            "type": "object",
            "properties": {
              "cwd": {
                "type": "string"
              },
              "root": {
                "type": "string"
              }
            },
            "required": [
              "cwd",
              "root"
            ]
          },
          "summary": {
            "type": "boolean"
          },
          "cost": {
            "type": "number"
          },
          "tokens": {
            "type": "object",
            "properties": {
              "input": {
                "type": "number"
              },
              "output": {
                "type": "number"
              },
              "reasoning": {
                "type": "number"
              },
              "cache": {
                "type": "object",
                "properties": {
                  "read": {
                    "type": "number"
                  },
                  "write": {
                    "type": "number"
                  }
                },
                "required": [
                  "read",
                  "write"
                ]
              }
            },
            "required": [
              "input",
              "output",
              "reasoning",
              "cache"
            ]
          },
          "finish": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "role",
          "time",
          "parentID",
          "modelID",
          "providerID",
          "mode",
          "agent",
          "path",
          "cost",
          "tokens"
        ]
      },
      "Message": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/UserMessage"
          },
          {
            "$ref": "#/components/schemas/AssistantMessage"
          }
        ]
      },
...
      "TextPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "text"
          },
          "text": {
            "type": "string"
          },
          "synthetic": {
            "type": "boolean"
          },
          "ignored": {
            "type": "boolean"
          },
          "time": {
            "type": "object",
            "properties": {
              "start": {
                "type": "number"
              },
              "end": {
                "type": "number"
              }
            },
            "required": [
              "start"
            ]
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "text"
        ]
      },
      "ReasoningPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "reasoning"
          },
          "text": {
            "type": "string"
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          },
          "time": {
            "type": "object",
            "properties": {
              "start": {
                "type": "number"
              },
              "end": {
                "type": "number"
              }
            },
            "required": [
              "start"
            ]
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "text",
          "time"
        ]
      },
...
      "FilePart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "file"
          },
          "mime": {
            "type": "string"
          },
          "filename": {
            "type": "string"
          },
          "url": {
            "type": "string"
          },
          "source": {
            "$ref": "#/components/schemas/FilePartSource"
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "mime",
          "url"
        ]
      },
...
      "ToolPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "tool"
          },
          "callID": {
            "type": "string"
          },
          "tool": {
            "type": "string"
          },
          "state": {
            "$ref": "#/components/schemas/ToolState"
          },
          "metadata": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type",
          "callID",
          "tool",
          "state"
        ]
      },
      "StepStartPart": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sessionID": {
            "type": "string"
          },
          "messageID": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "const": "step-start"
          },
          "snapshot": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "sessionID",
          "messageID",
          "type"
        ]
      },
...
      "Part": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/TextPart"
          },
          {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "sessionID": {
                "type": "string"
              },
              "messageID": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "const": "subtask"
              },
              "prompt": {
                "type": "string"
              },
              "description": {
                "type": "string"
              },
              "agent": {
                "type": "string"
              },
              "command": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "sessionID",
              "messageID",
              "type",
              "prompt",
              "description",
              "agent"
            ]
          },
          {
            "$ref": "#/components/schemas/ReasoningPart"
          },
          {
            "$ref": "#/components/schemas/FilePart"
          },
          {
            "$ref": "#/components/schemas/ToolPart"
          },
          {
            "$ref": "#/components/schemas/StepStartPart"
          },
          {
            "$ref": "#/components/schemas/StepFinishPart"
          },
          {
            "$ref": "#/components/schemas/SnapshotPart"
          },
          {
            "$ref": "#/components/schemas/PatchPart"
          },
          {
            "$ref": "#/components/schemas/AgentPart"
          },
          {
            "$ref": "#/components/schemas/RetryPart"
          },
          {
            "$ref": "#/components/schemas/CompactionPart"
          }
        ]
      },
...
      "Event": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/Event.tui.prompt.append"
          },
          {
            "$ref": "#/components/schemas/Event.tui.command.execute"
          },
          {
            "$ref": "#/components/schemas/Event.tui.toast.show"
          },
          {
            "$ref": "#/components/schemas/Event.tui.session.select"
          },
          {
            "$ref": "#/components/schemas/Event.installation.updated"
          },
          {
            "$ref": "#/components/schemas/Event.installation.update-available"
          },
          {
            "$ref": "#/components/schemas/Event.project.updated"
          },
          {
            "$ref": "#/components/schemas/Event.server.instance.disposed"
          },
          {
            "$ref": "#/components/schemas/Event.file.edited"
          },
          {
            "$ref": "#/components/schemas/Event.lsp.client.diagnostics"
          },
          {
            "$ref": "#/components/schemas/Event.permission.asked"
          },
          {
            "$ref": "#/components/schemas/Event.permission.replied"
          },
          {
            "$ref": "#/components/schemas/Event.session.status"
          },
          {
            "$ref": "#/components/schemas/Event.session.idle"
          },
          {
            "$ref": "#/components/schemas/Event.question.asked"
          },
          {
            "$ref": "#/components/schemas/Event.question.replied"
          },
          {
            "$ref": "#/components/schemas/Event.question.rejected"
          },
          {
            "$ref": "#/components/schemas/Event.todo.updated"
          },
          {
            "$ref": "#/components/schemas/Event.pty.created"
          },
          {
            "$ref": "#/components/schemas/Event.pty.updated"
          },
          {
            "$ref": "#/components/schemas/Event.pty.exited"
          },
          {
            "$ref": "#/components/schemas/Event.pty.deleted"
          },
          {
            "$ref": "#/components/schemas/Event.mcp.tools.changed"
          },
          {
            "$ref": "#/components/schemas/Event.file.watcher.updated"
          },
          {
            "$ref": "#/components/schemas/Event.lsp.updated"
          },
          {
            "$ref": "#/components/schemas/Event.command.executed"
          },
          {
            "$ref": "#/components/schemas/Event.vcs.branch.updated"
          },
          {
            "$ref": "#/components/schemas/Event.message.updated"
          },
          {
            "$ref": "#/components/schemas/Event.message.removed"
          },
          {
            "$ref": "#/components/schemas/Event.message.part.updated"
          },
          {
            "$ref": "#/components/schemas/Event.message.part.removed"
          },
          {
            "$ref": "#/components/schemas/Event.session.compacted"
          },
          {
            "$ref": "#/components/schemas/Event.session.created"
          },
          {
            "$ref": "#/components/schemas/Event.session.updated"
          },
          {
            "$ref": "#/components/schemas/Event.session.deleted"
          },
          {
            "$ref": "#/components/schemas/Event.session.diff"
          },
          {
            "$ref": "#/components/schemas/Event.session.error"
          },
          {
            "$ref": "#/components/schemas/Event.server.connected"
          },
          {
            "$ref": "#/components/schemas/Event.global.disposed"
          }
        ]
      },
...
      "BadRequestError": {
        "type": "object",
        "properties": {
          "data": {},
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "propertyNames": {
                "type": "string"
              },
              "additionalProperties": {}
            }
          },
          "success": {
            "type": "boolean",
            "const": false
          }
        },
        "required": [
          "data",
          "errors",
          "success"
        ]
      },
      "NotFoundError": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "const": "NotFoundError"
          },
          "data": {
            "type": "object",
            "properties": {
              "message": {
                "type": "string"
              }
            },
            "required": [
              "message"
            ]
          }
        },
        "required": [
          "name",
          "data"
        ]
      },
...
    }
  }
}